+++ cf_xarray/__init__.py
@@ -1,1 +1,2 @@
 from .accessor import CFAccessor  # noqa
+from .helpers import bounds_to_vertices, vertices_to_bounds  # noqa